feat: add IO panel with JSON import
