+++ src/stores/output.js
@@ -97,10 +97,12 @@ export const useOutputStore = defineStore('output', {
         },
         exportToJSON() {
             const { input } = useStore();
-            return `{
-                "input": { "src": "${input.src || ''}", "size": { "w": ${input.width || 0}, "h": ${input.height || 0} } },
-                "state": ${this.currentSnap()}
-            }`;
+            const state = JSON.parse(this.currentSnap());
+            delete state.history;
+            return JSON.stringify({
+                input: { src: input.src || '', size: { w: input.width || 0, h: input.height || 0 } },
+                state
+            });
         },
         exportToSVG() {
             const { nodeTree, nodes, pixels, viewport } = useStore();
